Added files via upload
This is my first upload for HW1 (excel-challenge). I have submitted the 'StarterBook (Moon)' XLSX file and a 'StarterBook_Report (Moon)' DOCX file. The Excel file is the altered Excel portion of the homework, the Word file is the written portion. The bonus challenges were included and labeled as such per their sheets in Excel file and as questions 4 and 5 in the Word file.